Create Bash Concat for Linux & Update .gitignore
The goal here is to be able to build this on Windows or Linux.  For
now the easiest thing seems to have both a bash and powershell scripts
for the same thing.

I have also created a `.gitignore` file and added the `public-db.sql`
artifact to it's ignore to prevent it from being committed into git
by mistake.