Added 21,479 mobs from Yggdrasil
This is a brute force copy over of mobs that appear to be missing in Eve that were in Yggdrasil.  Some regions, mostly Atlantis, dungeons, and housing regions, were completely empty and should now be populated.

I only copied over mobs with names that didn't appear in Eve, and I checked for duplicates as I went, so we shouldn't see a sudden increase in number of mobs that were already present.

Because of the sheer scope, I haven't fully tested this.  I set PackageID to Yggdrasil before copying mobs over, so if mobs do cause problems, it should be easy to remove them.