Fixed Paladin chants
The paladin dmg add chant was using the celerity chant as it's subspell, leading to the dmg chant having a 24/28/31/37% celerity buff when it should be 3/6/10/15%.  It also made the pallie celerity chant completely redundant.

I added correct celerity subspells and assigned them to the pallie dmg add chants.

I noticed that not all celerity buffs had an effectgroup, so I set them all to use effectgroup 1057 to keep them from stacking.  I'm not sure code enforces that, but at least the DB supports it.