Pre-emptively remove seal collector mobs
Removes dreaded seal mobs that might already be in the DB prior to creating ones using the DreadedSealCollector class.

There's a superfluous Lady Nina mob already present that I tried to remove from the Mob table,, but for some reason GitHub Desktop keeps uploading an empty file instead of the modified Mob.0.json.  This should resolve the issue in a manner that I can actually push.