Added
Makes items significantly more usable cross realm.  Items which would end up with duplicate bonuses are excluded.

I put the query removing class restrictions right at the top to make it easier to comment out.

To future proof this, queries looking for duplicate bonuses are included, to make it obvious when an item needs to be fixed by hand or added to the exclusion lists.